Avoid crashes on *nix systems. Fix #4.
It appears that cachemoney was developed on Windows and that
Windows and *nix likely have a different "key" format.

This patch avoided crashes such as #4 for me on linux, but I wouldn't be
surprised if it breaks on Windows. I'm not motivated to investigate this
further at the moment but I thought I'd push this as a reference for
what changes would need to be made to support *nix and a workaround for
*nix folks in the meantime.